@@ -7,8 +7,8 @@
     "Felix Edelmann <fxedel@gmail.com>"
   ],
   "engines": {
-    "node": "16.x",
-    "npm": "8.x"
+    "node": "20.x",
+    "npm": "10.x"
   },
   "type": "module",
   "scripts": {
@@ -25,7 +25,7 @@
     "build:api-docs": "widdershins --code true --language_tabs --omitBody true --omitHeader true --resolve true --outfile docs/rest-api.md ui/api/openapi.json",
 
     "clean": "rm -rf .nuxt tmp",
-    "update-package-lock": "rm package-lock.json && npm install && npm audit fix && npx browserslist@latest --update-db",
+    "update-package-lock": "rm package-lock.json && npm install && npm audit fix && npx update-browserslist-db@latest",
 
     "lint": "npm run lint:eslint && npm run lint:stylelint",
     "lint:eslint": "eslint .",
@@ -43,59 +43,63 @@
     "@nuxtjs/axios": "^5.13.6",
     "@nuxtjs/robots": "^2.5.0",
     "@nuxtjs/sitemap": "^2.4.0",
-    "@octokit/rest": "~19.0.13",
-    "a11y-dialog": "^7.5.2",
-    "ajv": "^8.12.0",
-    "ajv-formats": "^2.1.1",
-    "chalk": "^5.3.0",
+    "@octokit/rest": "~21.1.1",
+    "a11y-dialog": "^8.1.3",
+    "ajv": "^8.17.1",
+    "ajv-formats": "^3.0.1",
+    "chalk": "^5.4.1",
     "color-hash": "^2.0.2",
-    "color-name-list": "^10.10.0",
+    "color-name-list": "^10.28.1",
     "cookie-universal-nuxt": "^2.2.2",
     "cors": "^2.8.5",
     "embetty-vue": "^2.0.1",
-    "express": "^4.18.2",
+    "express": "^4.21.2",
     "jszip": "^3.10.1",
     "node-env-file": "^0.1.8",
-    "nuxt": "~2.17.0",
+    "nuxt": "~2.18.1",
     "nuxt-helmet": "^3.0.0",
-    "openapi-backend": "^5.9.2",
+    "openapi-backend": "^5.12.0",
     "sanitize-filename": "^1.6.3",
-    "sass": "^1.64.2",
-    "sass-loader": "^10.4.1",
+    "sass": "^1.86.3",
+    "sass-loader": "^10.5.2",
     "scroll-into-view": "^1.16.2",
     "svg-inline-loader": "^0.8.2",
-    "uuid": "^9.0.0",
+    "uuid": "^10.0.0",
     "vue-form": "^4.10.3",
-    "vue-ts-types": "^1.6.0",
-    "vue2-smooth-scroll": "^1.6.0",
+    "vue-ts-types": "^1.8.0",
     "vuedraggable": "^2.24.3",
     "xml2js": "^0.6.2",
     "xmlbuilder": "^15.1.1"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.2.8",
+    "@eslint/js": "~9.24.0",
+    "@eslint/markdown": "~6.4.0",
+    "@stylistic/stylelint-config": "~2.0.0",
     "broken-link-checker": "^0.7.8",
-    "dir-compare": "^4.0.0",
+    "dir-compare": "^5.0.0",
     "disparity": "^3.2.0",
-    "eslint": "~8.46.0",
-    "eslint-plugin-import": "~2.27.5",
-    "eslint-plugin-jsdoc": "~46.2.6",
-    "eslint-plugin-jsonc": "~2.9.0",
-    "eslint-plugin-markdown": "~3.0.1",
+    "eslint": "~9.24.0",
+    "eslint-plugin-import": "~2.31.0",
+    "eslint-plugin-jsdoc": "~50.0.1",
+    "eslint-plugin-jsonc": "~2.20.0",
     "eslint-plugin-nuxt": "~4.0.0",
-    "eslint-plugin-promise": "~6.1.1",
-    "eslint-plugin-sonarjs": "~0.19.0",
-    "eslint-plugin-unicorn": "~48.0.1",
-    "eslint-plugin-vue": "~9.15.1",
-    "eslint-plugin-vuejs-accessibility": "~2.1.0",
-    "jsdoc-to-markdown": "^8.0.0",
-    "libxmljs": "^1.0.9",
+    "eslint-plugin-promise": "~7.2.1",
+    "eslint-plugin-sonarjs": "~1.0.4",
+    "eslint-plugin-unicorn": "~58.0.0",
+    "eslint-plugin-vue": "~10.0.0",
+    "eslint-plugin-vuejs-accessibility": "~2.4.1",
+    "globals": "^16.0.0",
+    "jsdoc-to-markdown": "^9.1.1",
+    "libxmljs": "^1.0.11",
     "minimist": "^1.2.8",
-    "postcss-html": "^1.5.0",
-    "stylelint": "~15.10.2",
+    "postcss-html": "^1.8.0",
+    "stylelint": "~16.18.0",
     "stylelint-config-html": "~1.1.0",
-    "stylelint-config-recess-order": "~4.3.0",
-    "stylelint-config-standard-scss": "~10.0.0",
-    "stylelint-stylistic": "~0.4.3",
+    "stylelint-config-recess-order": "~6.0.0",
+    "stylelint-config-standard-scss": "~14.0.0",
+    "terminate": "^2.6.1",
+    "vue-eslint-parser": "~10.1.3",
     "widdershins": "^4.0.1"
   },
   "homepage": "https://open-fixture-library.org/",
