@@ -246,14 +246,13 @@ import { objectProp } from 'vue-ts-types';
 import { capabilityTypes, channelProperties } from '../../../lib/schema-properties.js';
 import {
   constants,
-  getEmptyFormState,
   getEmptyCapability,
   getEmptyFineChannel,
+  getEmptyFormState,
   getSanitizedChannel,
-  isChannelChanged,
   isCapabilityChanged,
-  clone,
-} from '../../assets/scripts/editor-utils.js';
+  isChannelChanged,
+} from '../../assets/scripts/editor-utilities.js';
 
 import A11yDialog from '../A11yDialog.vue';
 import LabeledInput from '../LabeledInput.vue';
@@ -275,12 +274,14 @@ export default {
     PropertyInputSelect,
     PropertyInputText,
   },
-  model: {
-    prop: `channel`,
-  },
   props: {
     channel: objectProp().required,
     fixture: objectProp().required,
+  },
+  emits: {
+    'channel-changed': () => true,
+    'remove-channel': channelId => true,
+    'reset-channel': () => true,
   },
   data() {
     return {
@@ -426,7 +427,7 @@ export default {
 
     copyPropertiesFromChannel(channel) {
       for (const property of Object.keys(channel)) {
-        this.channel[property] = clone(channel[property]);
+        this.channel[property] = structuredClone(channel[property]);
       }
     },
 
@@ -474,7 +475,7 @@ export default {
         WheelShake: /\bshake\b/i,
         WheelSlotRotation: /gobo ?\d* (?:rotation|index)/i,
         WheelRotation: /wheels? ?\d* (?:rotation|index)/i,
-        WheelSlot: /wheel|dis[ck]|gobos? ?\d*$/i,
+        WheelSlot: /wheel|dis[ck]|(?:gobos? ?\d*$)/i,
         EffectSpeed: /^(?:effect|program|macro) speed$/i,
         EffectDuration: /^(?:effect|program|macro) (?:time|duration)$/i,
         SoundSensitivity: /^(?:sound|mic|microphone) sensitivity$/i,
